feat: always sample at fixed, high rate
You are forced to give memsparkline a low `--wait` value
in order to avoid missing a brief memory usage spike.
This commit decouples the sampling rate, now fixed at 10 ms,
from the adjustable rate
at which memory usage is recorded and reported.
What is recorded and reported is the highest sample for the interval.

An extra check has been added to not miss the last record.

The sparkline with `minimum == maximum` is now `█████`,
not middle bars.
This is consistent with the sparkline for a single value