feat(cli): add `-r`/`--record` and `-s`/`--sample`
On my machine, the CPU usage with 25 ms sampling
is over 50% of one core.
This kind of load on the CPU
(and the corresponding battery drain on mobile devices)
is hard to justify as a default.
It is better to let the user determine the sampling rate.

Let's choose 200 ms as the new default sampling interval.
This reduces the load to around 10% of a core on the same machine.
memsparkline should still catch most of the same spikes at 200 ms.
In programs that run multiple seconds,
large memory spikes usually seem to last longer and/or repeat.